[Refactor]ElementsGroupコンポーネント実装
